Minor bugs are fixed without fixing http delete
