+++ WebUI/WebUI/settings.py
@@ -23,7 +23,8 @@ SECRET_KEY = '2lr1sdqkdl==*-na*%v@ky3k$e8y(qvv5tg)kb!!eu!71ix6y*'
 
 # SECURITY WARNING: don't run with debug turned on in production!
 # https://stackoverflow.com/questions/48276444/django-css-file-not-loading-in-production-debug-false/49398021
-DEBUG = False
+# DEBUG = False is used to show 404 error page
+DEBUG = True
 
 ALLOWED_HOSTS = ['*']
