+++ WebUI/static/css/page404.css
@@ -1,18 +1,10 @@
-.center404{
-    margin-left: 100px;
-    margin-right: 100px;
-    border: 3px solid #041E42;
-    padding: 10px;
-    text-align: center;
-}
-
-.center404 > a{
+.center > a{
     text-decoration: none;
     font-weight: bold;
     color: #041E42;
     font-size: medium;
 }
-.center404 > h1 > a{
+.center > h1 > a{
     text-decoration: none;
     font-weight: bold;
     color: red;
