+++ WebUI/templates/404.html
@@ -2,7 +2,7 @@
 {% load static %}
 {% block title %}MaSim LIMS / Page Error{% endblock %}
 {% block content %}
-<div class="center404">
+<div class="center">
   <h1><a href="/">404 Page Error: the page is missing or does not exist</a></h1>
   <a href="/"> Click Here: Go to Home Page </a>
 </div>
